+++ packages/connect-kit-loader/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Changed
+- The rpc parameter is now optional.
+- The chainId parameter is now optional, defaults to 1.
+
 ## 1.0.0-beta.4 - 2022-11-01
 
 ### Changed